requirements.txt
pennylane
matplotlib
scikit-learn
torch
scikit-image
seaborn
nibabel
ipykernel
imbalanced-learn
torchvision
hilbertcurve


streamlit>=1.28.0
torch>=2.0.0
pennylane>=0.32.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0 